requirements.txt
PyMuPDF==1.25.5
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.4.0
psycopg2-binary==2.9.12
pgvector==0.3.2
SQLAlchemy==2.0.31
ollama==0.3.3
celery==5.4.0
redis==5.0.8
kombu==5.3.7
google-generativeai==0.7.2
python-dotenv==1.0.1
pydantic==2.8.2
numpy>=1.26.4